# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API dla systemu ośrodka nadmorskiego
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [public]
      summary: Szczegóły pokoju
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Pokój
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Nie znaleziono
...
  /admin/rooms:
    get:
      operationId: adminListRooms
      tags: [admin]
      summary: Lista wszystkich pokoi (admin)
      responses:
        "200":
          description: Pokoje
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: adminCreateRoom
      tags: [admin]
      summary: Dodaj nowy pokój
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Nowy pokój
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /admin/rooms/{id}:
    get:
      operationId: adminGetRoom
      tags: [admin]
      summary: Szczegóły pokoju (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Pokój
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Nie znaleziono
    put:
      operationId: adminUpdateRoom
      tags: [admin]
      summary: Zaktualizuj pokój
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "200":
          description: Zaktualizowany pokój
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: adminDeleteRoom
      tags: [admin]
      summary: Usuń pokój
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Usunięto
  /admin/rooms/{id}/photos:
    post:
      operationId: adminAddRoomPhoto
      tags: [admin]
      summary: Dodaj zdjęcie do pokoju
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Zdjęcie dodane
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomPhoto"
  /admin/rooms/{id}/photos/{photoId}:
    delete:
      operationId: adminDeleteRoomPhoto
      tags: [admin]
      summary: Usuń zdjęcie pokoju
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
        - name: photoId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Usunięto

  # ─── ADMIN GALLERY ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Room:
      type: object
      required:
        - id
        - name
        - capacity
        - pricePerNight
        - isActive
        - sortOrder
        - amenities
        - photos
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        capacity:
          type: number
        pricePerNight:
          type: number
        minNights:
          type: number
        isActive:
          type: boolean
        sortOrder:
          type: number
        coverPhotoUrl:
          type: ["string", "null"]
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            $ref: "#/components/schemas/RoomPhoto"
    RoomInput:
      type: object
      required: [name, capacity, pricePerNight]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        capacity:
          type: number
          minimum: 1
        pricePerNight:
          type: number
          minimum: 0
        minNights:
          type: number
          minimum: 1
        isActive:
          type: boolean
        sortOrder:
          type: number
        coverPhotoUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
    RoomPhoto:
      type: object
      required: [id, roomId, url, sortOrder]
      properties:
        id:
          type: number
        roomId:
          type: number
        url:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: number
    PhotoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          minLength: 1
        caption:
          type: string

    # ─── GALLERY ─────────────────────────────────────────────────────────────
...