# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API dla systemu ośrodka nadmorskiego
 * OpenAPI spec version: 0.1.0
 */
import type { RoomPhoto } from './roomPhoto';

export interface Room {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  capacity: number;
  pricePerNight: number;
  minNights?: number;
  isActive: boolean;
  sortOrder: number;
  /** @nullable */
  coverPhotoUrl?: string | null;
  amenities: string[];
  photos: RoomPhoto[];
}